Fix timing of demand process and two additional minor issues.
This commit resolves three issues:

(1) The timing assumptions that we made for the demand process in the Matlab code slightly differed from what's in the paper. C_{t+1} is not known when firms make entry/exit decisions; neither to firms nor to the econometrician. Therefore, this commit updates the timing in the data generating process and in the likelihood.
(2) The timing of the cost shock W_t is also inconsistent with the paper. This is of course irrelevant, because W_t is iid. This commit changes the Matlab code such that the entry and survival decision in period t (i.e. N_{t+1}) are functions of W_t (not W_T)
(3) The initial values for the first step estimator are nonsensical, because there based on indices not values of C. This is now resolved, though it does not have any impact on the results.